+++ v2kit/models/base.py
@@ -55,7 +55,7 @@ class BaseConfig(ABC):
     @property
     def extra(self) -> Dict[str, object]:
         """Get extra data."""
-        return self._extra
+        return self._extra.copy()
 
     def update_label(
         self,
@@ -87,7 +87,7 @@ class BaseConfig(ABC):
             "Extra",
         )
 
-        self._extra = extra
+        self._extra = extra.copy()
 
         return self