+++ v2kit/utils.py
@@ -4,7 +4,7 @@ import base64
 from .errors import V2kitParseError
 from .validators import _validate_uri
 from .params import DEFAULT_ENCODING
-from .params import Protocol
+from .params import Protocol, SCHEME_TO_PROTOCOL
 
 
 def _encode_base64(data: str) -> str:
@@ -48,9 +48,9 @@ def _get_protocol(uri: str) -> Protocol:
     """
     _validate_uri(uri)
 
-    protocol = uri.split("://", 1)[0]
+    scheme = uri.split("://", 1)[0]
 
-    return Protocol(protocol)
+    return SCHEME_TO_PROTOCOL[scheme]
 
 
 def _is_protocol(uri: str, protocol: Protocol) -> bool: