+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `HttpConfig` class
 ### Changed
 - `README.md` updated
+- Test system modified
 ## [0.3] - 2026-06-16
 ### Added
 - Logo