+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `encoded_label` property
+### Changed
+- `parse` function modified
+- Test system modified
 ## [0.4] - 2026-07-02
 ### Added
 - `HttpConfig` class