+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `parse` function modified
 - `extra` property modified
 - `update_extra` method modified
+- `Protocol` enum updated
+- Scheme bug fixed
 - Test system modified
 ## [0.4] - 2026-07-02
 ### Added