+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `_get_protocol` function modified
+- `decode_subscription` function blank line bug fixed
 ## [0.5] - 2026-07-31
 ### Added
 - `encoded_label` property