+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - `_get_protocol` function modified
 - `decode_subscription` function blank line bug fixed
+- Port validation bug fixed
 ## [0.5] - 2026-07-31
 ### Added
 - `encoded_label` property