+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6] - 2026-09-01
 ### Added
 - `clear_extra` method
 - `set_extra_item` method
@@ -77,7 +78,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `encode_subscription` function
 - `decode_subscription` function
 
-[Unreleased]: https://github.com/sepandhaghighi/v2kit/compare/v0.5...dev
+[Unreleased]: https://github.com/sepandhaghighi/v2kit/compare/v0.6...dev
+[0.6]: https://github.com/sepandhaghighi/v2kit/compare/v0.5...v0.6
 [0.5]: https://github.com/sepandhaghighi/v2kit/compare/v0.4...v0.5
 [0.4]: https://github.com/sepandhaghighi/v2kit/compare/v0.3...v0.4
 [0.3]: https://github.com/sepandhaghighi/v2kit/compare/v0.2...v0.3